+++ node-llvm/constant.cpp
@@ -78,9 +78,9 @@ namespace jsllvm {
     else if (args.Length() == 3 && args[2]->IsNumber() && ty->getPrimitiveSizeInBits() == 64) {
       // allow a 3 arg form for 64 bit ints:
       // constant = llvm.Constant.getIntegerValue types.int64, ch, cl
-      int64_t vhi = v;
-      int64_t vlo = (int64_t)args[2]->NumberValue();
-      result = Value::New(llvm::Constant::getIntegerValue(ty, llvm::APInt(ty->getPrimitiveSizeInBits(), (int64_t)(vhi << 32 | vlo))));
+      uint64_t vhi = v;
+      uint32_t vlo = (uint32_t)args[2]->NumberValue();
+      result = Value::New (llvm::Constant::getIntegerValue(ty, llvm::APInt(ty->getPrimitiveSizeInBits(), (int64_t)((vhi << 32) | vlo))));
     }
     else {
       return ThrowException(Exception::TypeError(String::New("Invalid parameters to Constant.getIntegerValue")));
